Bookstore/Bookstore/requirements.txt
fastapi==0.75.0
uvicorn==0.17.0
pymysql==1.0.2
sqlalchemy==1.4.32
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.20.0